ci: fix Windows build after PR 2632 (#2660)
Fix the errors in Windows build after #2632:

- `kvdb/inverted_index_range_by_key_query_test.cpp`: explicit conversion from `std::filesystem::path` to `std::string`
- `snapshots/elias_fano/list_iterator.hpp`: `ListIterator::operator*` now returns `value_type` to fix MSVC conversion error. We could rename `ListIterator` as `ValueListIterator` to highlight it indeed iterates over primitive type values (i.e. integers)

plus other errors come to light after these:

- `local_transaction.cpp`, `concat_view_test.cpp`: add dual overloaded `operator==` and `operator!=` for `std::default_sentinel_t` in `ConcatView`

*Extras*
fix some MSVC warnings

+++ silkworm/db/datastore/snapshots/bittorrent/torrent_file.cpp
@@ -33,7 +33,7 @@
 
 namespace silkworm::snapshots::bittorrent {
 
-static constexpr int kDefaultPieceSize = 2_Mebi;
+static constexpr int kDefaultPieceSize = static_cast<int>(2_Mebi);
 
 TorrentFile::TorrentFile(ByteView data)
     : params_(lt::load_torrent_buffer(byte_view_to_str_span(data))) {
